Classes refactoring.
